jd - split webcontroller into HomeController and EarthquakesCOntroller and update HomepageTest
