+++ src/test/java/hello/HomePageTest.java
@@ -1,5 +1,8 @@
 package hello;
 
+import hello.controllers.HomeController;
+import hello.controllers.EarthquakesController;
+
 // import static org.hamcrest.Matchers.equalTo;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -21,7 +24,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(WebController.class)
+@WebMvcTest(HomeController.class)
 public class HomePageTest {
 
     @Autowired
